+++ tests/bench/bench_dict_reader.mojo
@@ -63,7 +63,7 @@ fn main():
     var time_single_small: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_dict_reader_small_single]()
-        time_single_small += elapsed / 1000000
+        time_single_small += Float64(elapsed) / 1000000
     var avg_single_small = time_single_small / 10
     print("Small Single-threaded:", round(avg_single_small, 4), "ms")
 
@@ -71,7 +71,7 @@ fn main():
     var time_multi_small: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_dict_reader_small]()
-        time_multi_small += elapsed / 1000000
+        time_multi_small += Float64(elapsed) / 1000000
     var avg_multi_small = time_multi_small / 10
     print("Small Threaded:", round(avg_multi_small, 4), "ms")
 
@@ -81,7 +81,7 @@ fn main():
     var time_single_medium: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_dict_reader_medium]()
-        time_single_medium += elapsed / 1000000
+        time_single_medium += Float64(elapsed) / 1000000
     var avg_single_medium = time_single_medium / 10
     print("Medium:", round(avg_single_medium, 2), "ms")
 
@@ -91,6 +91,6 @@ fn main():
     var time_single_large: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_dict_reader_large]()
-        time_single_large += elapsed / 1000000
+        time_single_large += Float64(elapsed) / 1000000
     var avg_single_large = time_single_large / 10
     print("Large:", round(avg_single_large, 1), "ms")
