+++ tests/bench/bench_threaded_diff.mojo
@@ -71,15 +71,15 @@ fn main():
     var time_single_small: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_single_threaded_small]()
-        time_single_small += elapsed / 1000000
+        time_single_small += Float64(elapsed) / 1000000
     var avg_single_small = time_single_small / 10
     print("Average time:", round(avg_single_small, 4), "ms")
-
+    
     print("Multi-threaded:")
     var time_multi_small: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_multi_threaded_small]()
-        time_multi_small += elapsed / 1000000
+        time_multi_small += Float64(elapsed) / 1000000
     var avg_multi_small = time_multi_small / 10
     print("Average time:", round(avg_multi_small, 4), "ms")
 
@@ -93,15 +93,15 @@ fn main():
     var time_single_medium: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_single_threaded_medium]()
-        time_single_medium += elapsed / 1000000
+        time_single_medium += Float64(elapsed) / 1000000
     var avg_single_medium = time_single_medium / 10
     print("Average time:", round(avg_single_medium, 2), "ms")
-
+    
     print("Multi-threaded:")
     var time_multi_medium: Float64 = 0
     for _ in range(10):
         var elapsed = time_function[bench_multi_threaded_medium]()
-        time_multi_medium += elapsed / 1000000
+        time_multi_medium += Float64(elapsed) / 1000000
     var avg_multi_medium = time_multi_medium / 10
     print("Average time:", round(avg_multi_medium, 2), "ms")
 
@@ -115,15 +115,15 @@ fn main():
     var time_single_large: Float64 = 0
     for _ in range(5):  # Fewer iterations for large file
         var elapsed = time_function[bench_single_threaded_large]()
-        time_single_large += elapsed / 1000000
+        time_single_large += Float64(elapsed) / 1000000
     var avg_single_large = time_single_large / 5
     print("Average time:", round(avg_single_large, 1), "ms")
-
+    
     print("Multi-threaded:")
     var time_multi_large: Float64 = 0
     for _ in range(5):  # Fewer iterations for large file
         var elapsed = time_function[bench_multi_threaded_large]()
-        time_multi_large += elapsed / 1000000
+        time_multi_large += Float64(elapsed) / 1000000
     var avg_multi_large = time_multi_large / 5
     print("Average time:", round(avg_multi_large, 1), "ms")
